Rename rules.json to rules_template.json
rules.json must contain the users investor ID and their authentication
token. We won't want these to be upstreamed. Rename the file to
rules_template.json. Users will need to copy this to rules.json and
update it accordingly. rules.json is in .gitignore so changes won't be
picked up.