Update how we deal with funding the account
Previously the account was only funded after good loans were found. Further more, we only transfered the order amount when the balance plus current transfers was less than the min_balance. This could result in needing two transfers to reach the min_balance. The new behavior transfers money if needed regardless of whether we find good loans to invest in. Transfers the smallest multiple of the orderamnt to reach the min_balance.